requirements.txt
pandas
python-telegram-bot
requests
numpy
matplotlib
